@@ -2,7 +2,7 @@
 require File.expand_path('../../core/lib/refinery/version', __FILE__)
 
 version = Refinery::Version.to_s
-rails_version = ['>= 3.1.11', '< 3.3'] | 0.upto(11).map{|i| "!= 3.2.#{i}"}
+rails_version = ['>= 3.1.11', '< 4.1']
 
 Gem::Specification.new do |s|
   s.platform          = Gem::Platform::RUBY
@@ -20,16 +20,18 @@ Gem::Specification.new do |s|
   s.files             = `git ls-files`.split("\n")
   s.test_files        = `git ls-files -- spec/*`.split("\n")
 
-  s.required_ruby_version = '>= 1.9.3'
+  s.required_ruby_version = '>= 2.0.0'
 
-  s.add_dependency 'refinerycms-i18n',            '~> 2.1.0'
-  s.add_dependency 'awesome_nested_set',          '~> 2.1.3'
+  s.add_dependency 'refinerycms-i18n',            '~> 3.0.0.dev'
+  s.add_dependency 'awesome_nested_set',          '~> 3.0.0.rc.1'
   s.add_dependency 'railties',                    rails_version
   s.add_dependency 'activerecord',                rails_version
   s.add_dependency 'actionpack',                  rails_version
   s.add_dependency 'truncate_html',               '~> 0.9'
   s.add_dependency 'will_paginate',               '~> 3.0.2'
-  s.add_dependency 'sass-rails',                  '~> 3.2.3'
-  s.add_dependency 'jquery-rails',                ['>= 2.0.0', '< 2.3.0']
+  s.add_dependency 'sass-rails',                  '~> 4.0.0'
+  s.add_dependency 'coffee-rails',                '~> 4.0.0'
+  s.add_dependency 'jquery-rails',                '>= 2.3.0'
+  s.add_dependency 'jquery-ui-rails',             '~> 4.0.0'
   s.add_dependency 'decorators',                  '~> 1.0.0'
 end
