@@ -5,33 +5,35 @@ module Refinery
   class User < Refinery::Core::BaseModel
     extend FriendlyId
 
-    has_and_belongs_to_many :roles, :join_table => :refinery_roles_users
+    has_and_belongs_to_many :roles, join_table: :refinery_roles_users
 
-    has_many :plugins, :class_name => "UserPlugin", :order => "position ASC", :dependent => :destroy
-    friendly_id :username, :use => [:slugged]
+    has_many :plugins, -> { order('position ASC') },
+                       class_name: "UserPlugin", dependent: :destroy
+
+    friendly_id :username, use: [:slugged]
 
     # Include default devise modules. Others available are:
     # :token_authenticatable, :confirmable, :lockable and :timeoutable
     if self.respond_to?(:devise)
       devise :database_authenticatable, :registerable, :recoverable, :rememberable,
-             :trackable, :validatable, :authentication_keys => [:login]
+             :trackable, :validatable, authentication_keys: [:login]
     end
 
     # Setup accessible (or protected) attributes for your model
     # :login is a virtual attribute for authenticating by either username or email
     # This is in addition to a real persisted field like 'username'
     attr_accessor :login
-    attr_accessible :email, :password, :password_confirmation, :remember_me, :username, :plugins, :login
+    attr_accessible :email, :password, :password_confirmation, :remember_me, :username, :plugins, :login, :full_name
 
-    validates :username, :presence => true, :uniqueness => true
-    before_validation :downcase_username
+    validates :username, presence: true, uniqueness: true
+    before_validation :downcase_username, :strip_username
 
     class << self
       # Find user by email or username.
       # https://github.com/plataformatec/devise/wiki/How-To:-Allow-users-to-sign_in-using-their-username-or-email-address
       def find_for_database_authentication(conditions)
         value = conditions[authentication_keys.first]
-        where(["username = :value OR email = :value", { :value => value }]).first
+        where(["username = :value OR email = :value", { value: value }]).first
       end
     end
 
@@ -46,7 +48,7 @@ module Refinery
           plugins.create(:name => plugin_name, :position => index)
         end
       else
-        assigned_plugins = plugins.all
+        assigned_plugins = plugins.load
         assigned_plugins.each do |assigned_plugin|
           if plugin_names.include?(assigned_plugin.name)
             plugin_names.delete(assigned_plugin.name)
@@ -56,8 +58,10 @@ module Refinery
         end
 
         plugin_names.each do |plugin_name|
-          plugins.create(:name => plugin_name,
-                         :position => plugins.pluck(:position).map(&:to_i).max + 1)
+          if plugin_name.is_a?(String)
+            plugins.create name: plugin_name,
+                           position: plugins.select(:position).map{|p| p.position.to_i}.max + 1
+          end
         end
       end
     end
@@ -104,7 +108,7 @@ module Refinery
     end
 
     def to_s
-      username.to_s
+      (full_name.presence || username).to_s
     end
 
     private
@@ -113,6 +117,12 @@ module Refinery
     # SELECT 1 FROM "refinery_users" WHERE LOWER("refinery_users"."username") = LOWER('UsErNAME') LIMIT 1
     def downcase_username
       self.username = self.username.downcase if self.username?
+    end
+
+    # To ensure that we aren't creating "admin" and "admin " as the same thing.
+    # Also ensures that "admin user" and "admin    user" are the same thing.
+    def strip_username
+      self.username = self.username.strip.gsub(/\ {2,}/, ' ') if self.username?
     end
 
   end
