@@ -16,7 +16,7 @@ module Refinery
       @user = User.new(params[:user])
 
       if @user.create_first
-        flash[:message] = "<h2>#{t('welcome', :scope => 'refinery.users.create', :who => @user.username).gsub(/\.$/, '')}.</h2>".html_safe
+        flash[:message] = "<h2>#{t('welcome', :scope => 'refinery.users.create', :who => @user.to_s).gsub(/\.$/, '')}.</h2>".html_safe
 
         sign_in(@user)
         redirect_back_or_default(refinery.admin_root_path)
