@@ -7,19 +7,23 @@ module Refinery
       def link_to
         # Get the switch_local variable to determine the locale we're currently editing
         # Set up Globalize with our current locale
-        Globalize.locale = params[:switch_locale] || Refinery::I18n.default_locale
+        Globalize.locale = if params[:switch_locale].present? && Refinery::I18n.built_in_locales.keys.map(&:to_s).include?(params[:switch_locale])
+          Globalize.locale = params[:switch_locale]
+        else
+          Refinery::I18n.default_locale
+        end
 
         @pages = ::Refinery::Page.roots.paginate(:page => params[:page], :per_page => ::Refinery::Page.per_page(true))
 
         @pages = @pages.with_globalize
 
         if ::Refinery::Plugins.registered.names.include?('refinery_files')
-            @resources = Resource.paginate(:page => params[:resource_page], :per_page => Resource.per_page(true)).
-                                  order('created_at DESC')
+          @resources = Resource.paginate(:page => params[:resource_page], :per_page => Resource.per_page(true)).
+                                order('created_at DESC')
 
           # resource link
           if params[:current_link].present?
-           is_resource_link = params[:current_link].include?("/system/resources")
+            is_resource_link = params[:current_link].include?("/system/resources")
           end
         end
 
