@@ -12,15 +12,24 @@ module Refinery
       end
 
       def template_options(template_type, current_page)
-        return {} if current_page.send(template_type)
+        html_options = { :selected => send("default_#{template_type}", current_page) }
 
-        if current_page.parent_id?
-          # Use Parent Template by default.
-          { :selected => current_page.parent.send(template_type) }
-        else
-          # Use Default Template (First in whitelist)
-          { :selected => Refinery::Pages.send("#{template_type}_whitelist").first }
+        if (template = current_page.send(template_type).presence)
+          html_options.update :selected => template
+        elsif current_page.parent_id? && !current_page.send(template_type).presence
+          template = current_page.parent.send(template_type).presence
+          html_options.update :selected => template if template
         end
+
+        html_options
+      end
+
+      def default_view_template(current_page)
+        current_page.link_url == "/" ? "home" : "show"
+      end
+
+      def default_layout_template(current_page)
+        "application"
       end
 
       # In the admin area we use a slightly different title
