@@ -46,7 +46,7 @@ end
 
 ActiveSupport.on_load(:active_record) do
   require 'awesome_nested_set'
-  require 'globalize3'
+  require 'globalize'
 end
 require 'friendly_id'
 require 'seo_meta'
