@@ -22,6 +22,7 @@ module Refinery
           expect(menu_presenter.selected_css).to eq(:selected)
           expect(menu_presenter.first_css).to eq(:first)
           expect(menu_presenter.last_css).to eq(:last)
+          expect(menu_presenter.list_tag_css).to eq("nav")
         end
       end
 
@@ -49,7 +50,7 @@ module Refinery
           menu_items = Refinery::Menu.new(FactoryGirl.create(:page, :title => "Refinery CMS"))
 
           menu_presenter = MenuPresenter.new(menu_items, view)
-          expect(menu_presenter.to_html).to eq(%Q{<nav class="menu clearfix" id="menu"><ul><li class="first last"><a href="/refinery-cms">Refinery CMS</a></li></ul></nav>})
+          expect(menu_presenter.to_html).to eq(%Q{<nav class="menu clearfix" id="menu"><ul class="nav"><li class="first last"><a href="/refinery-cms">Refinery CMS</a></li></ul></nav>})
         end
       end
 
